Renames to Contact and creates ContactForm component, moves form. Minor refactoring.
