+++ src/App.jsx
@@ -3,7 +3,7 @@ import LandingPage from './Components/LandingPage'
 import Projects from './Components/Projects'
 import AboutMe from './Components/AboutMe'
 import Sidebar from './Components/Sidebar'
-import ContactForm from './Components/ContactForm'
+import Contact from './Components/Contact'
 import Footer from './Components/Footer'
 import { Route } from 'react-router-dom'
 
@@ -14,7 +14,7 @@ const App = () => {
       <Route exact path="/" component={LandingPage} />
       <Route exact path="/projects" component={Projects} />
       <Route exact path="/about" component={AboutMe} />
-      <Route exact path="/contact" component={ContactForm} />
+      <Route exact path="/contact" component={Contact} />
       <Footer />
     </>
   )
